requirements.txt
accelerate==0.12.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cmake==3.26.1
contourpy==1.0.5
cycler==0.11.0
datasets==2.11.0
debugpy==1.6.3
decorator==5.1.1
deepspeed==0.8.3
defusedxml==0.7.1
diffusers==0.14.0
dill==0.3.5.1
einops==0.4.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
filelock==3.8.0
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.0
graphviz==0.20.1
hjson==3.1.0
huggingface-hub==0.13.4
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
install==1.3.5
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.14.0
jupyter-client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kiwisolver==1.4.4
lit==16.0.0
lpips==0.1.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.13
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
ninja==1.10.2.3
notebook==6.4.12
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyarrow==9.0.0
pycparser==2.21
pydantic==1.10.1
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.8.17
requests==2.28.1
responses==0.18.0
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.5.0
sympy==1.11.1
terminado==0.15.0
threadpoolctl==3.1.0
timm==0.6.7
tinycss2==1.1.1
tokenizers==0.12.1
torch==2.0.0
torch-fidelity==0.3.0
torchinfo==1.7.1
torchmetrics==0.9.3
torchsummary==1.5.1
torchvision==0.15.1
torchviz==0.0.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.27.4
triton==2.0.0
typing-extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.2
xxhash==3.0.0
yapf==0.32.0
yarl==1.8.1
zipp==3.8.1
